+++ src/app/App.tsx
@@ -1,17 +1,21 @@
+import { Suspense, lazy } from 'react';
 import { useSelector } from 'react-redux';
 import { Route, Routes } from 'react-router-dom';
 
-import { Layout, PrivateRoute, PublicRoute } from '@/index';
-import { RootState } from '@stores/store';
-import theme from '@styles/theme';
-import { Suspense, lazy } from 'react';
 import { ThemeProvider } from 'styled-components';
+
+import { RootState } from '@stores/store';
+
+import theme from '@styles/theme';
+
+import { Layout, PrivateRoute, PublicRoute } from '@/index';
+import { Center } from '@/pages/Center';
 
 function App() {
   const Login = lazy(() => import('@/pages/Login'));
   const Home = lazy(() => import('@/pages/Home'));
   const Me = lazy(() => import('@/pages/Me'));
-  const Ticket = lazy(() => import('@/pages/Ticket'));
+  const Sample = lazy(() => import('@/pages/Sample'));
 
   const isLogin = useSelector((state: RootState) => state.tokens.isLogin);
 
@@ -26,8 +30,9 @@ function App() {
               </Route>
               <Route element={<PrivateRoute isLogin={isLogin} />}>
                 <Route element={<Home />} path="/" />
-                <Route element={<Ticket />} path="tickets/*" />
+                <Route element={<Sample />} path="sample/*" />
                 <Route element={<Me />} path="me" />
+                <Route element={<Center />} path="Center/*" />
               </Route>
             </Route>
           </Routes>
